requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-JWT==0.3.2
Werkzeug==1.0.1
cython==0.29.21
pandas==1.2.4
numpy==1.20.2
firebase-admin==4.5.1
google-cloud-core==1.5.0
google-cloud-storage==1.35.0
google-cloud-vision==2.1.0
opencv-python-headless==4.5.1.48
Pillow==8.0.1
pytesseract==0.3.7
googlemaps==4.4.2
gunicorn==20.1.0
